backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

pydantic
python-jose
PyJWT
pdfplumber
python-docx
requests
